get rid of the AppRoutingModule
